refactor: update dependencies for improved performance and compatibility, and broken commands
-feature: news commands for jump and help

- Removed outdated dependencies: @discord-player/extractor, @discordjs/opus, @discordjs/voice, discord-player-youtubei
- Added @distube/ytdl-core and @snazzah/davey for enhanced functionality

+++ commands/stop.js
@@ -1,13 +1,37 @@
 import { SlashCommandBuilder } from 'discord.js';
 import { useQueue } from 'discord-player';
+import { noCurrentTrack } from '../bot-responses/errors.js';
 
 export const command = new SlashCommandBuilder()
-    .setName('stop') // Command name
+    .setName('stop')
     .setDescription('Para a música que está tocando atualmente');
 
 export async function execute(interaction) {
-
     const queue = useQueue(interaction.guild.id);
-    queue.delete();
 
+    if (!queue || !queue.currentTrack) {
+        return interaction.reply({
+            content: noCurrentTrack,
+            ephemeral: true,
+        });
+    }
+
+    const memberVoiceChannel = interaction.member.voice.channel;
+    if (!memberVoiceChannel || memberVoiceChannel.id !== queue.channel.id) {
+        return interaction.reply({
+            content: needToBeInSameVoiceChannel,
+            ephemeral: true,
+        });
+    }
+
+    try {
+        queue.delete();
+        await interaction.reply(stopResponse);
+    } catch (error) {
+        console.error(error);
+        await interaction.reply({
+            content: stoppingTrackError,
+            ephemeral: true,
+        });
+    }
 }
